refactor: httpRequest 리팩토링
