+++ src/main/java/request/HttpRequestUtils.java
@@ -1,10 +1,9 @@
-package webserver;
+package request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.UnsupportedEncodingException;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
